+++ packages/zen-persistent/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sylphx/zen-persistent
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/zen@2.0.0
+
 ## 2.0.0
 
 ### Patch Changes