+++ packages/zen-svelte/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sylphx/zen-svelte
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/zen@2.0.0
+
 ## 1.0.1
 
 ### Patch Changes